move bulk upload to magilink component and add email sending
