+++ src/Profile.jsx
@@ -3,7 +3,6 @@ import Promote from './Promote'
 import MagicLink from './MagicLink'
 import NavbarVhx from './Navbar'
 import Compute from './Compute'
-import BulkUpload from './BulkUpload'
 export default class Profile extends Component{
     constructor(props){
         super(props)
@@ -15,7 +14,6 @@ export default class Profile extends Component{
                 <Promote/>
                 <MagicLink/>
                 <Compute/>
-                <BulkUpload/>
             </div>
         )
     }
